+++ packages/keyring-api/src/events.ts
@@ -59,6 +59,16 @@ export const AccountCreatedEventStruct = object({
      * **Note:** This is not guaranteed to be honored by the MetaMask client.
      */
     displayConfirmation: exactOptional(boolean()),
+
+    /**
+     * Instructs MetaMask to display the name confirmation dialog in the UI.
+     * Otherwise, the account will be added with the suggested name, if it's not
+     * already in use; if it is, a suffix will be appended to the name to make it
+     * unique.
+     *
+     * **Note:** This is not guaranteed to be honored by the MetaMask client.
+     */
+    displayAccountNameSuggestion: exactOptional(boolean()),
   }),
 });