+++ packages/keyring-eth-ledger-bridge/jest.config.js
@@ -23,10 +23,10 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 90.9,
+      branches: 90.78,
       functions: 96,
-      lines: 95.05,
-      statements: 95.1,
+      lines: 95.02,
+      statements: 95.07,
     },
   },
 });