+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -31,7 +31,7 @@ import type {
 } from '@metamask/keyring-api';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 import { KeyringInternalSnapClient } from '@metamask/keyring-internal-snap-client';
-import type { JsonRpcRequest } from '@metamask/keyring-utils';
+import type { AccountId, JsonRpcRequest } from '@metamask/keyring-utils';
 import { strictMask } from '@metamask/keyring-utils';
 import type { SnapId } from '@metamask/snaps-sdk';
 import { type Snap } from '@metamask/snaps-utils';
@@ -101,8 +101,10 @@ export type SnapKeyringCallbacks = {
     address: string,
     snapId: SnapId,
     handleUserInput: (accepted: boolean) => Promise<void>,
+    onceSaved: Promise<AccountId>,
     accountNameSuggestion?: string,
     displayConfirmation?: boolean,
+    displayAccountNameSuggestion?: boolean,
   ): Promise<void>;
 
   removeAccount(
@@ -205,7 +207,13 @@ export class SnapKeyring extends EventEmitter {
       account: newAccountFromEvent,
       accountNameSuggestion,
       displayConfirmation,
+      displayAccountNameSuggestion,
     } = message.params;
+
+    // READ THIS CAREFULLY:
+    // ------------------------------------------------------------------------
+    // The account creation flow is now asynchronous. We expect the Snap to
+    // first create the account data and then fire the "AccountCreated" event.
 
     // Potentially migrate the account.
     const account = transformAccount(newAccountFromEvent);
@@ -224,18 +232,66 @@ export class SnapKeyring extends EventEmitter {
       throw new Error(`Account '${account.id}' already exists`);
     }
 
+    // A deferred promise that will be resolved once the Snap keyring has saved
+    // its internal state.
+    // This part of the flow is run asynchronously, so we have no other way of
+    // letting the MetaMask client that this "save" has been run.
+    // NOTE: Another way of fixing that could be to rely on events through the
+    // messenger maybe?
+    const onceSaved = new DeferredPromise<AccountId>();
+
+    // Add the account to the keyring, but wait for the MetaMask client to
+    // approve the account creation first.
     await this.#callbacks.addAccount(
       address,
       snapId,
+      // This callback is passed to the MetaMask client, it will be called whenever
+      // the end user will accept or not the account creation.
       async (accepted: boolean) => {
         if (accepted) {
+          // We consider the account to be created on the Snap keyring only if
+          // the user accepted it. Meaning that the Snap MIGHT HAVE created the
+          // account on its own state, but the Snap keyring MIGHT NOT HAVE it yet.
+          //
+          // e.g The account creation dialog crashed on MetaMask, this callback
+          // will never be called, but the Snap still has the account.
           this.#accounts.set(account.id, { account, snapId });
-          await this.#callbacks.saveState();
+
+          // This is the "true async part". We do not `await` for this call, mainly
+          // because this callback will persist the account on the client side
+          // (through the `AccountsController`).
+          //
+          // Since this will happen after the Snap account creation and Snap
+          // event, if anything goes wrong, we will delete the account by
+          // calling `deleteAccount` on the Snap.
+          // eslint-disable-next-line no-void
+          void this.#callbacks
+            .saveState()
+            .then(() => {
+              // This allows the MetaMask client to be "notified" when then
+              // Snap keyring has truly persisted its state. From there, we should
+              // be able to use the account (e.g. to display account creation
+              // confirmation dialogs).
+              onceSaved.resolve(account.id);
+            })
+            .catch(async (error) => {
+              // FIXME: There's a potential race condition here, if the Snap did
+              // not persist the account yet (this should mostly be for older Snaps).
+              await this.#deleteAccount(snapId, account);
+
+              // This allows the MetaMask client to be "notified" that something went
+              // wrong with the Snap keyring. (e.g. useful to display account creation
+              // error dialogs).
+              onceSaved.reject(error);
+            });
         }
       },
+      onceSaved.promise,
       accountNameSuggestion,
       displayConfirmation,
+      displayAccountNameSuggestion,
     );
+
     return null;
   }
 
@@ -1095,6 +1151,16 @@ export class SnapKeyring extends EventEmitter {
   async removeAccount(address: string): Promise<void> {
     const { account, snapId } = this.#resolveAddress(address);
 
+    await this.#deleteAccount(snapId, account);
+  }
+
+  /**
+   * Removes an account.
+   *
+   * @param snapId - Snap ID.
+   * @param account - Account to delete.
+   */
+  async #deleteAccount(snapId: SnapId, account: KeyringAccount): Promise<void> {
     // Always remove the account from the maps, even if the Snap is going to
     // fail to delete it.
     this.#accounts.delete(snapId, account.id);
@@ -1106,7 +1172,7 @@ export class SnapKeyring extends EventEmitter {
       // with the account deletion, otherwise the account will be stuck in the
       // keyring.
       console.error(
-        `Account '${address}' may not have been removed from snap '${snapId}':`,
+        `Account '${account.address}' may not have been removed from snap '${snapId}':`,
         error,
       );
     }