+++ packages/keyring-eth-trezor/CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING:** `TrezorKeyring` now implements the `Keyring` type ([#194](https://github.com/MetaMask/accounts/pull/194))
+  - The class does not extend `EventEmitter` anymore.
+  - The `TrezorKeyring.accounts` class variable is now a `readonly Hex[]` array.
+  - The `addAccounts` method signature has been changed:
+    - An `amount` number parameter is now required to specify the number of accounts to add.
+    - The method now returns a promise resolving to an array of `Hex` addresses.
+  - The `deserialize` method now requires a `TrezorControllerOptions` object as a parameter.
+  - The `unlock` method now returns `Promise<Hex>`.
+  - The `getAccounts` method now returns `Promise<Hex[]>`.
+  - The `signTransaction` method now accepts an `Hex` typed value as the `address` parameter.
+  - The `signMessage` method now accepts an `Hex` typed value as the `withAccount` parameter.
+  - The `signPersonalMessage` method now accepts an `Hex` typed value as the `withAccount` parameter.
+  - The `signTypedData` method now accepts an `Hex` typed value as the `withAccount` parameter.
+  - The `unlockAccountByAddress` method now accepts an `Hex` typed value as the `address` parameter.
+
+### Removed
+
+- **BREAKING:** The `exportAccount` method has been removed ([#194](https://github.com/MetaMask/accounts/pull/194))
+
+## [6.1.1]
+
+### Changed
+
+- Use `ts-bridge/cli@0.6.3` ([#214](https://github.com/MetaMask/accounts/pull/214))
+  - This new version fixes a bug regarding some missing exports.
+
 ## [6.1.0]
 
 ### Added
@@ -161,7 +189,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support new versions of ethereumjs/tx ([#88](https://github.com/metamask/eth-trezor-keyring/pull/88))
 
-[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.1.1...HEAD
+[6.1.1]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.1.0...@metamask/eth-trezor-keyring@6.1.1
 [6.1.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.0.2...@metamask/eth-trezor-keyring@6.1.0
 [6.0.2]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.0.1...@metamask/eth-trezor-keyring@6.0.2
 [6.0.1]: https://github.com/MetaMask/accounts/compare/@metamask/eth-trezor-keyring@6.0.0...@metamask/eth-trezor-keyring@6.0.1